+++ docs/CHANGELOG.md
@@ -10,6 +10,14 @@
   - Add note about linking/using templates (html and latex) with pandoc.
   - Add references to argmap specs spreadsheet?
 
+## argmap 10.18.0
+
+- `test/test_scripts/bash_aliases_argmap_test.sh`: Update and add tests.
+  - `test/test_scripts/bash_aliases_argmap_test.sh`:
+    - `test_function()`
+    - `test_getvar()`
+    - `test_get_site_path()`
+
 ## argmap 10.17.0
 
 - `scripts/bash_aliases_mapjs.sh`: